pyproject.toml
[project]
name = "llm-router"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "modal",
    "petname>=2.6",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.3",
]

[project.scripts]
llama-router-worker = "llama_worker:main"
